requirements.txt
Flask
python-dotenv
requests
psycopg2
flask_login
flask_SQLAlchemy
email_validator